creating and deleting posts working

+++ app/Http/Controllers/AdminUsersController.php
@@ -164,7 +164,10 @@ class AdminUsersController extends Controller
     {
         
         $user = User::findOrFail($id);
-        unlink(public_path() . $user->photo->file);
+        if(isset($user->photo->file)) {
+            unlink(public_path() . $user->photo->file);
+        }
+        
         $user->delete();
 
         Session::flash('deleted_user', 'The user has been deleted');
